requirements.txt
anthropic>=0.34.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
gevent>=24.0.0
gunicorn>=22.0.0
